UnitCell

Chemfiles.UnitCell — Type

A UnitCell describes the bounding box of a system. It is represented by three base vectors of lengths a, b, and c; and the angles between these vectors are alpha, beta, and gamma.

Chemfiles.CellShape — Type

The possible shape for an unit cell are:

Chemfiles.Orthorhombic for unit cells with the three angles are 90°.

Chemfiles.Triclinic for unit cells where the three angles may not be 90°.

Chemfiles.Infinite for unit cells without boundaries.

Chemfiles.UnitCell — Method

UnitCell(a::Number, b::Number, c::Number, α::Number, β::Number, γ::Number)

Create an UnitCell from the three lenghts and three angles.

Chemfiles.UnitCell — Method

UnitCell(frame::Frame)

Get a copy of the UnitCell of a frame.

Chemfiles.UnitCell — Method

UnitCell(a::Number, b::Number, c::Number)

Create an UnitCell from the three lenghts, with all the angles equal to 90°.

Base.deepcopy — Method

deepcopy(cell::UnitCell)

Make a deep copy of a cell.

Chemfiles.angles — Method

angles(cell::UnitCell)

Get the three cell angles (alpha, beta, and gamma) in degrees.

Chemfiles.lengths — Method

lengths(cell::UnitCell)

Get the three cell lengths (a, b, and c) in angstroms.

Chemfiles.matrix — Method

matrix(cell::UnitCell)

Get the cell matricial representation, i.e. the representation of the three base vectors as::

    | a_x   b_x   c_x |
    |  0    b_y   c_y |
    |  0     0    c_z |

Chemfiles.set_angles! — Method

set_angles!(cell::UnitCell, α::Real, β::Real, γ::Real)

Set the cell angles to α, β, and γ.

α, β, and γ should be in degrees.

Chemfiles.set_lengths! — Method

set_lengths!(cell::UnitCell, a::Real, b::Real, c::Real)

Set the cell lengths to a, b, and c.

a, b, and c should be in angstroms.

Chemfiles.set_shape! — Method

set_shape!(cell::UnitCell, shape::CellShape)

Set the cell shape to the given shape.

Chemfiles.shape — Method

shape(cell::UnitCell)

Get the cell shape, as a CellShape value.

Chemfiles.volume — Method

volume(cell::UnitCell)

Get the unit cell volume.

Chemfiles.wrap! — Method

wrap!(cell::UnitCell, vector::Vector{Float64})

Wrap a vector in the unit cell.