Frame

Chemfiles.Frame — Type

A Frame holds data for one step of a simulation. As not all formats provide all the types of information, some fields may be initialized to a default value. A Frame may contain the following data:

Positions for all the atoms in the system;

Velocities for all the atoms in the system;

The Topology of the system;

The UnitCell of the system.

Chemfiles.Frame — Method

Frame() -> Frame

Create a new empty Frame.

Base.angle — Method

angle(frame::Frame, i::Integer, j::Integer, k::Integer) -> Float64

Calculate the angle made by three atoms.

Base.deepcopy — Method

deepcopy(frame::Frame) -> Frame

Make a deep copy of a Frame.

Base.length — Method

length(frame::Frame) -> Int64

Get the number of atoms in the frame.

Base.resize! — Method

resize!(frame::Frame, natoms::Integer)

Resize the positions and the velocities in the frame, to make space for natoms atoms. This function may invalidate any pointer to the positions or the velocities if the new size is bigger than the old one. In all the cases, previous data is conserved. This function conserve the presence or absence of velocities.

Base.size — Method

size(frame::Frame) -> Int64

Get the number of atoms in the frame.

Base.step — Method

step(frame::Frame) -> UInt64

Get the frame step, i.e. the frame number in the trajectory.

Chemfiles.add_atom! — Function

add_atom!(frame::Frame, atom::Atom, position::Array{Float64,1})
add_atom!(frame::Frame, atom::Atom, position::Array{Float64,1}, velocity::Array{Float64,1})

Add an atom and the corresponding position and velocity data to a frame.

Chemfiles.add_bond! — Function

add_bond!(frame::Frame, i::Integer, j::Integer)
add_bond!(frame::Frame, i::Integer, j::Integer, order)

Add an additional bond to the Frame's Topology.

Chemfiles.add_residue! — Method

add_residue!(frame::Frame, residue::Residue)

Add a residue to the Frame's Topology.

Chemfiles.add_velocities! — Method

add_velocities!(frame::Frame)

Add velocities to this frame. The storage is initialized with the result of size(frame) as the number of atoms. If the frame already has velocities, this does nothing.

Chemfiles.clear_bonds! — Method

clear_bonds!(frame::Frame)

Remove all bonds, angles and dihedral angles from the Frame's Topology.

Chemfiles.dihedral — Method

dihedral(frame::Frame, i::Integer, j::Integer, k::Integer, m::Integer) -> Float64

Calculate the dihedral (torsional) angle made by four unbranched atoms.

Chemfiles.distance — Method

distance(frame::Frame, i::Integer, j::Integer) -> Float64

Calculate the distance between two atoms.

Chemfiles.guess_bonds! — Method

guess_bonds!(frame::Frame)

Guess the bonds, angles, and dihedrals in the frame using a distance criteria.

Chemfiles.has_velocities — Method

has_velocities(frame::Frame) -> Bool

Check if a frame contains velocity data or not.

Chemfiles.list_properties — Method

list_properties(frame::Frame) -> Array{String,1}

Get the names of all properties associated with a frame.

Chemfiles.out_of_plane — Method

out_of_plane(frame::Frame, i::Integer, j::Integer, k::Integer, m::Integer) -> Float64

Calculate the out-of-plane (improper) angle made by four atoms.

Chemfiles.positions — Method

positions(frame::Frame) -> Chemfiles.ChemfilesArray

Get the positions in a Frame as an array. The positions are readable and writable from this array. If the frame is resized (by writing to it, or calling resize!), the array is invalidated.

Chemfiles.properties_count — Method

properties_count(frame::Frame) -> Int64

Get the number of properties associated with a frame.

Chemfiles.property — Method

property(frame::Frame, name::String) -> Any

Get a named property for the given atom.

Chemfiles.remove_atom! — Method

remove_atom!(frame::Frame, index::Integer)

Remove the atom at index from the frame.

This function modifies all the atoms indexes after index, and invalidates any array obtained using positions or velocities.

Chemfiles.remove_bond! — Method

remove_bond!(frame::Frame, i::Integer, j::Integer)

Remove a bond from the Frame's Topology.

Chemfiles.set_cell! — Method

set_cell!(frame::Frame, cell::UnitCell)

Set the cell associated with a frame.

Chemfiles.set_property! — Method

set_property!(frame::Frame, name::String, value)

Set a named property for the given Frame.

Chemfiles.set_step! — Method

set_step!(frame::Frame, step::Integer)

Set the frame step to step.

Chemfiles.set_topology! — Method

set_topology!(frame::Frame, topology::Topology)

Set the topology associated with a frame.

Chemfiles.velocities — Method

velocities(frame::Frame) -> Chemfiles.ChemfilesArray

Get the velocities in a Frame as an array. The velocities are readable and writable from this array. If the frame is resized (by writing to it, or calling resize!), the array is invalidated.

If the frame do not have velocity, this function will error. You can use add_velocities! to add velocities to a frame before calling this function.