Miscelaneous functions

Chemfiles.add_configuration — Method

Chemfile.add_configuration(path::String)

Read configuration data from the file at path.

By default, chemfiles reads configuration from any file name .chemfilesrc in the current directory or any parent directory. This function can be used to add data from another configuration file.

This function will fail if there is no file at path, or if the file is incorectly formatted. Data from the new configuration file will overwrite any existing data.

Chemfiles.clear_errors — Method

Chemfiles.clear_errors()

Clear any error messages stored by the chemfiles runtime.

Chemfiles.last_error — Method

Chemfiles.last_error()

Get the last error message from the chemfiles runtime.

Chemfiles.set_warning_callback — Method

Chemfiles.set_warning_callback(callback::Function)

Set the global warning callback to be used for each warning event.

The callback function must take a String and return nothing.