Residue

Chemfiles.Residue — Type

A Residue is a group of atoms belonging to the same logical unit. They can be small molecules, amino-acids in a protein, monomers in polymers, etc.

Chemfiles.Residue — Type

Residue(name::String) -> Residue
Residue(name::String, id) -> Residue

Create a new residue with the given name and optional residue identifier id.

Chemfiles.Residue — Method

Residue(topology::Topology, index::Integer) -> Residue

Get a copy of the residue at index from a topology.

The residue index in the topology is not always the same as the residue identifier.

Base.contains — Method

contains(residue::Residue, index::Integer) -> Bool

Check if the atom at the given index is in the residue.

Base.deepcopy — Method

deepcopy(residue::Residue) -> Residue

Make a deep copy of a residue.

Base.size — Method

size(residue::Residue) -> Int64

Get the number of atoms in a residue.

Chemfiles.add_atom! — Method

add_atom!(residue::Residue, index::Integer)

Add the atom at the given index in the residue.

Chemfiles.atoms — Method

atoms(residue::Residue) -> Vector{UInt64}

Get the atoms in a $residue$. This function returns a list of indexes.

Chemfiles.id — Method

id(residue::Residue) -> Int64

Get the identifier of a residue in the initial topology.

Chemfiles.list_properties — Method

list_properties(residue::Residue) -> Vector{String}

Get the names of all properties associated with a residue.

Chemfiles.name — Method

name(residue::Residue) -> String

Get the name of a residue.

Chemfiles.properties_count — Method

properties_count(residue::Residue) -> Int64

Get the number of properties associated with a residue.

Chemfiles.property — Method

property(residue::Residue, name::String) -> Any

Get a named property for the given residue.

Chemfiles.residue_for_atom — Method

residue_for_atom(
    topology::Topology,
    index::Integer
) -> Union{Nothing, Residue}

Get a copy of the residue containing the atom at index in the topology.

This function will return nothing if the atom is not in a residue, or if the index is bigger than the number of atoms in the topology.

Chemfiles.set_property! — Method

set_property!(
    residue::Residue,
    name::String,
    value::Union{Bool, Float64, String, Vector{Float64}}
)

Set a named property for the given residue.