Trajectory

Chemfiles.Trajectory — Type

A Trajectory represents a simulation file on the hard drive. It can read or write one or many Frames to this file. The file format can be automatically determined from the extention, or manually specified. Writing to a Trajectory is buffered, which means that one needs to close() the trajectory and flush the buffer before being able to read the file again.

Chemfiles.Trajectory — Type

Trajectory(path::AbstractString, mode::Char='r', format::AbstractString="")

The Trajectory function opens a trajectory file, using the file at the given path. The opening mode can be 'r' for read, 'w' for write or 'a' for append, and defaults to 'r'. The optional format parameter give a specific file format to use when opening the file.

Chemfiles.Trajectory — Method

Trajectory(f::Function, args...)

Apply the function f to the result of Trajectory(args...) and close the resulting trajectory upon completion, similar to open(f, args...).

Base.close — Method

close(trajectory::Trajectory)

Close a trajectory. This function flushes any buffer content to the hard drive, and frees the associated memory. Necessary when running on the REPL to finish writing.

Base.isopen — Method

isopen(trajectory::Trajectory)

Check if the trajectory is open.

Base.length — Method

length(trajectory::Trajectory)

Get the number of steps (the number of frames) in a trajectory.

Base.read — Method

read(trajectory::Trajectory)

Read the next step of the trajectory, and return the corresponding Frame.

Base.size — Method

size(trajectory::Trajectory)

Get the number of steps (the number of frames) in a trajectory.

Base.write — Method

write(trajectory::Trajectory, frame::Frame)

Write the given frame to the trajectory.

Chemfiles.path — Method

path(trajectory::Trajectory)

Get the path used to open a trajectory.

Chemfiles.read_step — Method

read_step(trajectory::Trajectory, step::Integer)

Read the given step of the trajectory, and return the corresponding Frame.

Chemfiles.set_cell! — Method

set_cell!(trajectory::Trajectory, cell::UnitCell)

Set the cell associated with a trajectory. This cell will be used when reading and writing the file, replacing any unit cell in the file.

Chemfiles.set_topology! — Function

set_topology!(trajectory::Trajectory, path::AbstractString, format::AbstractString = "")

Set the Topology associated with a trajectory by reading the first frame of the file at path; and extracting the topology of this frame. The optional format parameter can be used to specify the file format.

Chemfiles.set_topology! — Method

set_topology!(trajectory::Trajectory, topology::Topology)

Set the Topology associated with a trajectory. This topology will be used when reading and writing the file, replacing any topology in the file.