Selection

Chemfiles.Selection — Type

A Selection is used to select a group of atoms. Examples of selections are "name H" and "(x < 45 and name O) or name C". See the full documentation <http://chemfiles.org/chemfiles/latest/selections.html>_ for more information about the selection language.

Chemfiles.Selection — Method

Selection(selection::AbstractString) -> Selection

Create a Selection from a selection string.

Base.deepcopy — Method

deepcopy(selection::Selection) -> Selection

Make a deep copy of a selection.

Base.size — Method

size(selection::Selection) -> UInt64

Get the size of a selection, i.e. the number of atoms we are selecting together.

Chemfiles.evaluate — Method

evaluate(selection::Selection, frame::Frame) -> Vector{Any}

Evaluate a selection on a given frame. This function return a list of indexes or tuples of indexes of atoms in the frame matching the selection.

Chemfiles.selection_string — Method

selection_string(selection::Selection) -> String

Get the selection string used to create a given selection.