AtomsBase integration

Chemfiles is integrated with the AtomsBase by means of appropriate convert methods as well as respective constructors, which are listed below. Note that these conversion routines copy data unlike the majority of functionality in Chemfiles.

If your main interest is in reading structural data from disk, have a look at AtomsIO. This package provides a uniform interface for reading structural data and exposing them in an AtomsBase-compatible way. Under the hood the package employs a number of parser packages (including Chemfiles) to support reading / writing a wide range of file formats.

AtomsBase.AbstractSystem — Method

Construct an AtomsBase AbstractSystem from a Chemfiles Frame.

AtomsBase.FlexibleSystem — Method

Construct an AtomsBase FlexibleSystem from a Chemfiles Frame.

Base.convert — Method

convert(
    _::Type{AbstractSystem},
    frame::Frame
) -> FlexibleSystem

Convert a Chemfiles Frame to an AtomsBase AbstractSystem

Base.convert — Method

convert(
    _::Type{FlexibleSystem},
    frame::Frame
) -> FlexibleSystem

Convert a Chemfiles Frame to an AtomsBase FlexibleSystem