Miscelaneous functions

Chemfiles.FormatMetadata — Type

Metadata associated with one of the format Chemfiles can read/write

name::String

Name of the format

extension::Union{Nothing, String}

Extension associated with the format, or nothing if there is no extension associated.

description::String

Extended user-facing description of the format

reference::String

URL pointing to the format definition/reference

read::Bool

Is reading files in this format implemented?

write::Bool

Is writing files in this format implemented?

memory::Bool

Does this format support in-memory IO?

positions::Bool

Does this format support storing atomic positions?

velocities::Bool

Does this format support storing atomic velocities?

unit_cell::Bool

Does this format support storing unit cell information?

atoms::Bool

Does this format support storing atom names or types?

bonds::Bool

Does this format support storing bonds between atoms?

residues::Bool

Does this format support storing residues?

Chemfiles.add_configuration — Method

add_configuration(path::String)

Read configuration data from the file at path.

By default, chemfiles reads configuration from any file name .chemfilesrc in the current directory or any parent directory. This function can be used to add data from another configuration file.

This function will fail if there is no file at path, or if the file is incorectly formatted. Data from the new configuration file will overwrite any existing data.

Chemfiles.clear_errors — Method

clear_errors()

Clear any error messages stored by the chemfiles runtime.

Chemfiles.format_list — Method

format_list() -> Array{Chemfiles.FormatMetadata,1}

Get the full list of formats supported by Chemfiles, and associated metadata

Chemfiles.last_error — Method

last_error() -> String

Get the last error message from the chemfiles runtime.

Chemfiles.set_warning_callback — Method

set_warning_callback(callback::Function)

Set the global warning callback to be used for each warning event.

The callback function must take a String and return nothing.